Added parameter methodForObtainingIpAddress
I added a parameter methodForObtainingIpAddress allowed values of Dynamic and Static.  I built two variables one object ipaddressMethod with Static and Dynamic as branches that had the properties reflecting the settings required for the Network Interface resource.  Then I used a dot source method so the parameter would select the correct configuration.  "ipAddressObject": "[variables('ipAddressMethod')[parameters('methodForObtainingIpAddress')]]"  I added the variable ipAddressObject to the Network Interface resource to set the configuration as desired.  

So the IP address would be changed to static after obtaining an IP, I added a nested template with a reference method using the configured IP and changing the "privateIPAllocationMethod" setting to static.